backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
python-multipart>=0.0.5
aiohttp>=3.8.0
scikit-learn>=1.0.0
pytest>=7.0.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 